+++ pik/core/shortcuts/request.py
@@ -1,0 +1,8 @@
+from typing import Optional
+
+from django.http import HttpRequest
+from simple_history.models import HistoricalRecords
+
+
+def get_current_request() -> Optional[HttpRequest]:
+    return getattr(HistoricalRecords.thread, 'request', None)