+++ release.sh
@@ -31,3 +31,4 @@ echo "NEW VERSION: $VERSION"
 python setup.py sdist
 python -m twine upload dist/${NAME}-${VERSION}.tar.gz
 git tag -a v$VERSION -m "version $VERSION"
+git push --tags